Replace datetime with vec_datetime_value (#258)

+++ be/src/vec/data_types/data_type_date.h
@@ -21,7 +21,7 @@
 
 namespace doris::vectorized {
 
-class DataTypeDate final : public DataTypeNumberBase<Int128> {
+class DataTypeDate final : public DataTypeNumberBase<Int64> {
 public:
     TypeIndex get_type_id() const override { return TypeIndex::Date; }
     const char* get_family_name() const override { return "Date"; }
@@ -33,7 +33,7 @@ public:
     std::string to_string(const IColumn& column, size_t row_num) const;
     void to_string(const IColumn &column, size_t row_num, BufferWritable &ostr) const override;
 
-    static void cast_to_date(Int128& x);
+    static void cast_to_date(Int64& x);
 };
 
 } // namespace doris::vectorized
